move embeddng to chat folder
