+++ templates/types/streaming/nextjs/app/api/chat/embed/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { initSettings } from "../engine/settings";
-import { readAndSplitDocument } from "./embeddings";
+import { readAndSplitDocument } from "../embeddings";
 
 initSettings();
