+++ templates/types/streaming/nextjs/app/api/chat/embeddings.ts
